fix: RAM/VRAM precision to 2 decimals, remove TAP hint
- Remove 'TAP ANYWHERE TO EXIT' footer from system status screen
- Change RAM/VRAM display from 1 to 2 decimal places (%.2f GB)
- Fix RAM misalignment with htop: use float precision throughout
  the entire pipeline (host kB→MB, wire protocol, firmware parser)
  instead of truncated integers that lost ~0.5MB accuracy
- Update telemetry struct ram/vram fields from int to float
- Update host log format to show fractional MB values

+++ firmware/main/telemetry.h
@@ -9,12 +9,12 @@
 typedef struct {
     int64_t epoch_utc;
     int cpu_pct;
-    int ram_used_mb;
-    int ram_total_mb;
+    float ram_used_mb;
+    float ram_total_mb;
     int gpu_pct;
     int gpu_temp_c;
-    int vram_used_mb;
-    int vram_total_mb;
+    float vram_used_mb;
+    float vram_total_mb;
     int64_t last_received_us;
     bool has_data;
 } pet_telemetry_t;
